requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
editdistance==0.5.3
habanero==0.7.2
hammock==0.2.4
idna==2.9
lxml==4.5.0
numpy==1.18.2
orkg==0.4.2
pandas==1.0.2
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
tei-reader==0.0.17
termcolor==1.1.0
tqdm==4.43.0
urllib3==1.25.8
